pyproject.toml
[tool.poetry]
name = "sheet-sql"
version = "0.1.0"
description = "SQL for Google Sheets"
authors = ["Andy Huynh <andy.huynh312@gmail.com>"]
readme = "README.md"
repository = "github.com/andyh1203/sheet-sql"
packages = [
    { include = "sheetsql", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.24.0"
gspread = "^3.6.0"
regex = "^2020.6.8"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
black = "^19.10b0"
pre-commit = "^2.5.1"
mypy = "^0.780"
invoke = "^1.4.1"
mock = "^4.0.2"
sphinx = "^3.1.1"
nox = "^2020.5.24"
coverage = {extras = ["toml"], version = "^5.1"}
pytest-cov = "^2.10.0"
flake8-annotations = "^2.1.0"
flake8-black = "^0.2.0"
flake8-bugbear = "^20.1.4"
flake8-bandit = "^2.1.2"
flake8-docstrings = "^1.5.0"
safety = "^1.9.0"
pytype = {version = "^2020.6.26", python = "3.7"}

[tool.coverage.paths]
source = ["src", "*/site-packages"]

[tool.coverage.run]
branch = true
source = ["sheetsql"]

[tool.coverage.report]
show_missing = true
fail_under = 70

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
